requirements.txt
playsound
openai-whisper
sounddevice==0.5.1
numpy
SpeechRecognition
pygame
requests
elevenlabs==0.2.26
torch==2.6.0
torchaudio
gTTS
soundfile
sseclient
RPi.GPIO
flask
psutil
opencv-python
picamera2
smbus2
jinja2
Werkzeug
itsdangerous
click
pyserial
ultralytics
loguru
scipy==1.9.3
lap
cython_bbox
transformers==4.50.1
tqdm
deepface
pillow
sounddevice==0.5.1
tqdm
tf-keras
faster-whisper

